Minor changes to SingleTrackLearningInstructions
Mostly minor copyedits to the document. The only semantic change is that I believe where the document said "10 minutes" in one place, it should have said "5 minutes", like in the other places that a time limit is referenced. But please check this in case I misunderstood what this part of the instructions was referring to.